修改.gitignore； 增加停止播放按钮；

+++ app/src/main/java/demo/abooc/com/upnp/activity/VideoPlayerView.java
@@ -118,11 +118,11 @@ public class VideoPlayerView {
         mVolumeMute.setImageResource(mute ? R.drawable.ic_action_volume_muted : R.drawable.ic_action_volume_on);
     }
 
-    public void setVolume(int volume) {
+    public void setVolume(long volume) {
         mVolumeValue.setText("" + volume);
     }
 
-    public void seekVolume(int volume) {
-        mSeekBarVolume.setProgress(volume);
+    public void seekVolume(long volume) {
+        mSeekBarVolume.setProgress((int) volume);
     }
 }
